Добавил конвертер, исправил таймер (теперь срок его службы где-то на уровне жизни вселенной), протестил

+++ sketch_arduino/SendMessageRequest.h
@@ -6,7 +6,7 @@
 class SendMessageRequest{
   public:
 
-  SendMessageRequest(std::vector<byte> message, uint32_t send_time, bool software_serial){
+  SendMessageRequest(std::vector<byte> message, uint64_t send_time, bool software_serial){
     for (int i = 0; i < message.size(); i++){
       this->message = message;
     }
@@ -14,7 +14,7 @@ class SendMessageRequest{
     this->software_serial = software_serial;
   }
 
-  uint32_t get_send_time(){
+  uint64_t get_send_time(){
     return send_time;
   }
 
@@ -29,6 +29,6 @@ class SendMessageRequest{
   private:
   
   std::vector<byte> message;
-  uint32_t send_time;
+  uint64_t send_time;
   bool software_serial;
 };
